pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scrapix"
version = "0.1.0"
description = "Scrapix - Smart, fast, and simple image scraper for Google Search"
readme = "README.md"
license = { text = "MIT" }
authors = [
    {name = "Jonas Renault", email = "jonasrenault@gmail.com"},
]
requires-python = ">=3.12"
dependencies = [
    "fake-useragent>=2.2.0",
    "pillow>=12.0.0",
    "pydantic-settings>=2.11.0",
    "selenium>=4.37.0",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
]

[project.urls]
"Homepage" = "https://github.com/jonasrenault/scrapix#readme"
"Repository" = "https://github.com/jonasrenault/scrapix"

[project.scripts]
scrapix = "scrapix.cli.main:app"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.1",
]

[tool.black]
line_length = 90
target_version = ['py312']

[tool.ruff]
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
]
src = ["scrapix", "tests"]
line-length = 90

[tool.ruff.lint]
ignore = []
select = [
    "E",
    "F",
    "W",
    "I001",
]

[tool.ruff.lint.isort]
known-first-party = ["scrapix"]

[tool.mypy]
files = ["scrapix", "tests"]
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
    'ignore::DeprecationWarning',
]
